Fix circular dependencies in webrtc_common.
One reason for the circular deps is that common_types.h is a
historical dumping ground for various structs and defines that
are believed to be generally useful. I tried moving things out
that did not appear to be used downstream (StreamCounters,
RtpCounters etc) and moved the things that seemed used
(RtpHeader + supporting structs) to a new file api/rtp_headers.h.
This makes their place in the api more clear while moving out
the things that don't belong in the API in the first place.

I had to extract out typedefs.h from webrtc_common to resolve
another circular dependency. I believe checks includes typedefs,
but common depends on checks.

Bug:webrtc:7745

+++ system_wrappers/include/ntp_time.h
@@ -40,8 +40,12 @@ class NtpTime {
   // NTP standard (RFC1305, section 3.1) explicitly state value 0 is invalid.
   bool Valid() const { return value_ != 0; }
 
-  uint32_t seconds() const { return value_ / kFractionsPerSecond; }
-  uint32_t fractions() const { return value_ % kFractionsPerSecond; }
+  uint32_t seconds() const {
+    return static_cast<uint32_t>(value_ / kFractionsPerSecond);
+  }
+  uint32_t fractions() const {
+    return static_cast<uint32_t>(value_ % kFractionsPerSecond);
+  }
 
  private:
   uint64_t value_;
